Point: Alias descale method
Bit of a dirty fix for a funny "bug".

The bug in question is that running `Point.descale`, it throws the following error:

```
Decustom: taps must be greater than 0.
```

This is obviously because the kernel radius for Point is 0. But as descaling Point is equivalent to scaling with Point, we can simply work around this by aliasing the descale method with the scale method.