Remove `EwaBicubic`
Firstly, it was incorrectly referencing ewa_robidoux, which _is_ a
bicubic but internally its `b` and `c` params take precedence over
anything a user might have passed to `EwaBicubic`, so functionally it
was equivalent to `EwaRobidoux`. Secondly, there is no "ewa_bicubic" to
replace it with; there's "bicubic" but I think we've got that kernel
sufficiently covered with `Bicubic` elsewhere.